feat(web): add the staging Onboarding Lab route
Add the Web half of the staging-only Onboarding Lab at
`/internal/onboarding-lab`, plus the operator documentation for running it.

The route sits inside authenticated routing but outside the setup-completion
gate, so a tester can reopen it when a real onboarding run gets stuck after a
reset. The Server decides whether the Account may use the Lab, and an Account
that may not renders Not Found. A persistent warning states that the staging
Account is shared and that resetting it can interrupt another tester.

Scenario Preview renders fixed onboarding states through the production page
rather than a second one: the presentation moves out of `OnboardingPage` into
`OnboardingView`, which the page and the Lab both render, so Preview reuses the
real state derivation, hierarchy and copy. Preview writes nothing — it mounts no
Feishu setup lifecycle, and Agent creation reads no creation intent, resumes
nothing and takes no focus. Only the selected fixture is kept, in the URL.

The reset action asks for one confirmation, calls the internal interface, and on
a verified 204 refreshes authoritative `/me` state before entering the ordinary
`/onboarding` route with no stale Agent selection. A failed reset stays on the
Lab with a retry and never infers success from client state.

Refs #180
